add default route for drawer
