+++ packages/plugin-ethereum/client/src/index.ts
@@ -4,6 +4,7 @@ import Nodes from '@magickml/plugin-ethereum-shared'
 import Money from '@mui/icons-material/AttachMoney'
 
 import { EthereumComponent } from './components/ethereum.route'
+import { ContractComponent } from './components/contract.route'
 
 type StartEthereumArgs = {
   agent: any
@@ -16,15 +17,19 @@ const EthereumPlugin = new ClientPlugin({
   agentComponents: [EthereumAgentWindow],
   drawerItems: [
     {
-      path: '/ethereum/:chain/:address/:function',
+      path: '/ethereum/',
       icon: Money,
       text: 'Ethereum',
     },
   ],
   clientRoutes: [
     {
-      path: "/ethereum/:chain/:address/:function",
+      path: "/ethereum/",
       component: EthereumComponent,
+    },
+    {
+      path: "/ethereum/contract/:chain/:address/:function",
+      component: ContractComponent,
     },
   ],
   secrets: [
